WIP move DeleteUserDialog for own component
