+++ frontend/src/pages/CreateForm/QuestionList.jsx
@@ -5,7 +5,7 @@ import { useFieldArray, useFormContext } from 'react-hook-form';
 import Input from '../../components/Form/Input.jsx';
 import { SmallButton } from '../../components/Buttons/SmallButton.jsx';
 import { useCustomFormProvider } from '../../context/FormContext.jsx';
-import TitleLink from '../../components/Header/TitleLink.jsx';
+import TitleLink from '../../components/UserNavbar/TitleLink.jsx';
 
 export const QuestionList = () => {
   const { swapQuestion, addQuestion, register, getValues } = useCustomFormProvider();
