+++ frontend/src/pages/Workspace/Workspace.module.css
@@ -1,30 +1,14 @@
-.viewport {
-    margin: 0px;
-    height: 100vh;
-    background: rgb(160, 158, 194);
-    background: linear-gradient(90deg, rgba(160, 158, 194, 0.5410758053221288) 0%, rgba(197, 197, 255, 0.5550814075630253) 43%, rgba(0, 212, 255, 0.4710477941176471) 100%);
-    width: 100vw;
-    overflow-y: auto;
-}
-
-h1 {
-    margin-top: 10px;
-    margin-left: 20px;
-}
-
-.frame {
+/* .frame {
     display: grid;
     margin: 20px;
     background-color: azure;
     padding: 20px;
     border-radius: 20px;
-    grid-template-columns: 1fr 4fr;
+    grid-template-columns: 1fr 3fr;
+    gap: 20px;
     height: auto;
-    -webkit-box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
-    -moz-box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
     box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
-}
-
+} */
 
 .btn {
     font-family: "Space Mono", monospace;
